requirements-train.txt
torch
accelerate
peft @ git+https://github.com/huggingface/peft.git
trl
bitsandbytes
flash-attn
requests
autoawq